Corrected poll rate was cramping on my style.  I was already accounting for it, so accounting for it again made the speeds appear slower than they actually were.

+++ NetLive/src/main/java/com/richardlucasapps/netlive/UnitConverter.java
@@ -1,5 +1,5 @@
 package com.richardlucasapps.netlive;
 
 public interface UnitConverter {
-    public double convert(long bytesPerSecond);
+    public double convert(double bytesPerSecond);
 }
